requirements.txt
numpy>=1.21.0
RPi.GPIO>=0.7.0
pyserial>=3.5
smbus2>=0.4.1